pyproject.toml
[project]
name = "autoreclipper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "google-generativeai>=0.8.5",
    "loguru>=0.7.3",
    "pillow>=11.2.1",
    "pynput>=1.8.1",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.0",
    "pywin32>=310",
]
